refactor to allow for post processing of air quality metrics
