+++ server/air_quality/calculate.js
@@ -1,0 +1,5 @@
+function average(values, hours) {
+	return values.pop();
+}
+
+exports.average = average;
